% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{train_test_split}
\alias{train_test_split}
\title{Train Test Split Function}
\usage{
train_test_split(x, length, pr, data, numtr, numts, shapex, shapey)
}
\arguments{
\item{x}{Determines how many dummys there will be}

\item{length}{The your array or dataframe length.}

\item{pr}{Probablity for the new array.}

\item{data}{Your data.}

\item{numtr}{numtr is number dummy for train array.}

\item{numts}{numts is number dummy for test array.}

\item{shapex}{For shapex xtrain&xtest. Used to select rows or columns.}

\item{shapey}{For shapey ytrain&ytest. Used to select rows or columns.}
}
\description{
The traintest method divides the dataset into 2 parts for people working in the field of machine learning, and the train part is used to build the model and the test part is used to test the model.
}
\references{
By Eren Altuntas(2021)
}
\seealso{
The x argument must be typed.
}
